small changes
SoldierDiaryPerformanceState: button highlight instead of invis
TransferItemsState: append wounds to Soldier names
Map: hide smoke if tile is undiscovered
UnitInfoState: require RMB to exit (not LMB)
Scrollbar: FIXED !
Soldier::getName() use size_t instead of unsigned.

+++ src/Savegame/Soldier.h
@@ -141,7 +141,7 @@ private:
 		/// Gets the soldier's name.
 		std::wstring getName(
 				bool statstring = false,
-				unsigned int maxLength = 20) const;
+				size_t maxLength = 20) const;
 		/// Sets the soldier's name.
 		void setName(const std::wstring& name);
